backend/requirements.txt
django-ninja==1.*
django-stubs-ext
django-types
Django==5.*
gunicorn
psycopg2-binary==2.*
pydantic-settings==2.*
pydantic==2.*
whitenoise==6.*
